+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add support for `async`, `sync`, `sync*` functions
 - Add support for ann expression to `asAwait`, `asYield`, `asYieldStar`
+- Add `toExportBuilder` and `toImportBuilder` to types and references
 
 ## 1.0.0-alpha+7